pyproject.toml
[tool.poetry]
name = "pairing-matrix"
version = "0.1.0"
description = "Pair Programming Matrix"
authors = [
    "jochen preusche <iilei@users.noreply.github.com>",
]

license = "BSD-3-Clause"
readme = "README.md"

documentation = "https://github.com/iilei/pairing_matrix/blob/master/README.md"
homepage = "https://github.com/iilei/pairing_matrix"
repository = "https://github.com/iilei/pairing_matrix"

classifiers = [
  "Development Status :: 4 - Beta",
  "Environment :: Console",

  "Intended Audience :: Developers",
  "Operating System :: OS Independent",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.7",
  "Topic :: Software Development :: Libraries :: Python Modules",
]


[tool.poetry.dependencies]
python = "~=3.7"
PyGithub = "^1.53"
tabulate = "^0.8.7"
confuse = "^1.3.0"
python-dotenv = "^0.14.0"
pyyaml = "^5.3.1"
asyncio = "^3.4.3"
config42 = {git = "https://github.com/yurilaaziz/config42.git", rev = "116c4bca91f36858f8e70fdf5f5151bba8188b36"}
python-gitlab = "^2.5.0"
python-datemath = "^1.5.2"
pytz = "^2020.1"
ramda = "^0.5.7"
arrow = "^0.17.0"
gitpython = "^3.1.9"
rich = "^8.0.0"
deepmerge = "^0.1.0"
numpy = "^1.19.2"
pandas = "^1.1.3"


[tool.poetry.dev-dependencies]
black = "=19.3b0"
flake8 = "^3.7"
flake8-isort = "^2.7"
ipython = "^7.6"
isort = "^4.3"
mypy = "^0.711.0"
pdbpp = "^0.10.0"
pre-commit = "^1.17"
pytest = "^3.0"
pytest-black = "^0.3.7"
pytest-cov = "^2.7"

pytest-flake8 = "^1.0"
pytest-mock = "^1.10"
pytest-mypy = "^0.3.3"
pytest-randomly = "^3.0"
tox = "^3.13"


[tool.black]
line-length = 89
skip-string-normalization = true
target_version = ['py37']
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
)
'''


[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
